requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
requests==2.31.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
apscheduler==3.10.4
loguru==0.7.2
jinja2==3.1.2
python-multipart==0.0.6
openpyxl==3.1.2